refactor(eval): move judge modules to eval/judges/ subpackage
Moves 4 judge-related modules from eval/ root into a new eval/judges/
subpackage to reduce eval/ top-level clutter (~30 files → ~26):

- eval/llm_judge.py        → eval/judges/llm_judge.py
- eval/synthetic_judge.py  → eval/judges/synthetic_judge.py
- eval/judge_agreement.py  → eval/judges/judge_agreement.py
- eval/judge_common.py     → eval/judges/judge_common.py

Updates:
- eval/judges/__init__.py (new, makes it a regular package)
- Internal imports: from eval.judge_common → from eval.judges.judge_common
- ROOT_DIR / ROOT path depth: parents[1] → parents[2] (one extra level)
- External imports in tests/ (5 files) and scripts/llm_judge.py
- Makefile target paths (eval/llm_judge.py → eval/judges/llm_judge.py,
  -m eval.synthetic_judge → -m eval.judges.synthetic_judge)
- Sphinx :mod:`eval.judge_common` references in docstrings

eval/config.yaml stays at eval/ root (ADR 0001 naive_baseline preset
path constraint preserved).

ADR 0001 naive_baseline invariant + ADR 0003 answer dict contract:
unchanged (no retrieval/verifier behavior change).

84 judge tests pass; 11 pre-existing failures on main (harness +
mixed-format-ingestion) are unrelated.

Closes #705
